feat(pkg/tzap): use tzapio/tzap's custom ListFilesInDir function in LoadFileDir method
This commit changes the `LoadFileDir` method in the `pkg/tzap/file.go` file to use the custom `ListFilesInDir` function from `tzapio/tzap/pkg/util` instead of the standard library's `filepath.Glob`. This allows for more flexibility in file matching. The `match` parameter has been removed as it is no longer needed. The `panic` statement on error has been replaced with a `Printf` statement and another `panic` statement to reduce the likelihood of silent failures. The commit also updates the corresponding test file.