feat: enable logs by default
Changed the logic of the "enablelogs" flag to default to true, and renamed it to "disablelogs" for usability purposes. This change will ensure that logs are enabled by default and can be disabled using the new flag. This commit also updates the corresponding flag usage in the CLI's root command.