feat(cli): introduce refactor command
This commit introduces a new command `refactor` that allows for refactoring code either using command-line flags or using a configuration file. This command is used to generate refactored code or produce documentation files. The command makes use of the `tzap` package, and includes command-line flags for specifying `filein`, an optional `fileout`, the `task` to perform, a mission to describe the overall purpose of the project, an optional `plan` to steer output, `outputformat`, an optional `example` to provide an example for the refactoring task, and `inspiration` flag to specify inspiration files. Additionally, a JSON example has been provided to demonstrate how to use the command with a configuration file.