test(pkg/tzap): update file test to use new MkdirTemp function
Instead of using os.TempDir() to create temporary directories and files, the MkdirTemp function is now being used. This change ensures that temporary files and directories are created in the system's temporary directory with a name that begins with the provided prefix. This commit updates the file_test.go to use the new MkdirTemp function.