refactor: add support for dry run option
Add support for a dry run option in the refactor command. If the option is set to true, the command will not write to fileout, but will only print to stdout. This is useful for previewing the changes that would be made, before actually making them.

BREAKING CHANGE: N/A