feat(tzapconnect): Add DeleteEmbeddingDocuments method
This commit adds a new method called DeleteEmbeddingDocuments to the tzapconnect package. This method allows multiple embedding documents to be deleted at once by passing a slice of document IDs. The method is implemented in the connector.go file for the PartialComposite struct, and also in the StubConnector implementation. The TGenerator interface in types/tgenerator.go is updated to include the new method definition.

This change is a new feature, and thus is marked with the "feat" type in the commit message.