feat(cli): add embedding search workflow with confirmation and inspiration files
This commit adds a new workflow for searching embeddings with confirmation and inspiration files. The confirmation step has been added to the `LoadAndFetchEmbeddings` function. In this function, `ConfirmEmbeddingSearch` and `PrintInspirationFiles` have been refactored to be named workflows. The `embeddingPromptCmd` has been modified accordingly to use these named workflows.

The `PrintInspirationFiles` workflow prints the specified inspiration files to `stdout`. It also reads the number of tokens and prints that alongside the filename. The `ConfirmEmbeddingSearch` workflow checks if there are uncached embeddings and if the number of embeddings to fetch is more than 19. If true, it displays a prompt to the user asking if they want to proceed. The prompt also displays an estimate of the price.

In addition, `embeddingPromptCmd` has been modified to accept a search query as input from API mode. In this case, if there is no search query provided, the command will panic and prompt the user to enter a search query.

BREAKING CHANGE: This commit adds a breaking change since `embeddingPromptCmd` is modified to use `RootCmd` instead of `rootCmd`.