feat(cmdutil): Add FileInDirEvaluator to evaluate file inclusion/exclusion rules
This commit adds a new type `FileInDirEvaluator` which is responsible for evaluating which files and directories should be included in a given file tree, based on `.gitignore`, `.tzapignore`, and `.tzapinclude` files. It contains methods to traverse a directory using rules defined in an exclude list, and also get all the files in a tree while applying inclusion and exclusion rules. It allows users to provide custom inclusion/exclusion rules by passing arguments.

This is a new feature that will not affect backward compatibility, hence the `feat` type has been used.

BREAKING CHANGE: N/A.