feat(openaiconnector): update tokenizer to use new tokenizer package
BREAKING CHANGE: The OpenAI Connector now uses a new tokenizer package that is not compatible with the previous one used. This requires any existing code using the OpenAI Connector's tokenizer to be updated to use the new package.

Update the tokenizer used by the OpenAI Connector to the new `tzapio/tokenizer` package, specifically using the `cl100k_base` codec. The previous `tiktoken-go/tokenizer` package is no longer being used. This change includes updating the import statements, initializing the new codec as `_tokenizer`, and removing the old initialization code.

This is a breaking change as any code using the previous tokenizer package will need to be updated to use the new one.