feat(pkg/embed): improve caching of existing files
This commit improves file caching functionality in the `pkg/embed` package. It introduces a file cache utility that checks whether files that are about to be processed have changed by comparing MD5 hashes of the current files content with their previous hashes stored in `.tzap-data/filesmd5.db`. The files that have not changed are placed in a map called `unchangedFiles`, while the changed files are placed in a different map called `changedFiles`.

To achieve this improvement, some changes have also been made to the `PrepareEmbeddingsFromFiles` function to use the `changedFiles` map instead of the `files` slice. The `GetDrift` function has also been updated to not delete the embeddings for files that have not changed since the last time they were processed. Finally, the function `FetchAndCacheNewEmbeddings` has been updated to store the MD5 hash of a file content in the database, in addition to storing its embedding.

These significant changes should improve the performance and efficiency of the codebase, especially when processing large numbers of files. This commit should not break any existing functionality.