feat(util): add MD5HashByte method for byte array hashing
This commit adds a new method to the utils package, MD5HashByte, that takes a byte array as input and returns the corresponding MD5 hash as a hex-encoded string. This method can be used for hashing binary data, in contrast to the existing MD5Hash method which takes a string as input.