refactor: update default parameters for searchCmd
This commit updates the default values for the "embeds" and "ncount" flags for the searchCmd function. The values have been changed to match the default values used in the "embeddingprompt" library. The "embeds" value has been updated from 20 to 10 and the "ncount" value has been updated from 30 to 20. This change ensures that the searchCmd function is more in line with the default parameters used by the embeddingprompt library.