fix(pkg): return values in CheckFileCache function
The CheckFileCache function was not returning any values on error in the case of a localdb error. This commit updates the code to return nil values and the error in case the localdb cannot be created. Additionally, the function is updated to return the correct values in the case when there are no changed files.