refactor(export): move GetEmbeddingsFromFile function to export package
This commit refactors the `GetEmbeddingsFromFile` function from the `pkg/embed/embed.go` file and moves it to the `pkg/embed/export/export.go` file, as it is a more appropriate location considering its functionality of retrieving embeddings from a JSON file. The function is now exported and can be accessed outside of the package. The `json` package import has also been moved to this file, and a new import of the `tl` package has been added to handle logging. This commit does not include any breaking changes.