feat: add IsZero function to reflectutil package
This commit adds a new function called `IsZero` to the `reflectutil` package. This function takes an input of type interface{} and returns a boolean value indicating whether the input is zero or not. This function uses the reflect package to compare the input value against its zero value.

Also, a new test file called `zero_test.go` is added to test the `IsZero` function with various inputs.

BREAKING CHANGE: This change does not break any existing functionality.