feat(embed): add SingleWait type for synchronizing a long running process
This commit adds a new type `SingleWait` in the `pkg/embed/localdb/singlewait` package to synchronize a long running process. The `SingleWait` type has two methods: `New` and `GetData`. `New` creates a new instance of `SingleWait` and initiates the `Done` channel. The `Process` method simulates a long running process and closes the `Done` channel when finished. `GetData` waits for the `Process` method to finish and then returns the data.

BREAKING CHANGE: This commit adds a new feature to the codebase.