feat(embed): add Go gob encoding and decoding functions
This commit adds two new functions, `NewGobReaderIO` and `NewGobWriterIO`, which provide a way to easily serialize and deserialize Go structures using gob encoding and decoding. This will be particularly useful for local databases that need to store data in a binary format that can easily be read and written.

The `NewGobReaderIO` function takes an `io.Reader` as a parameter and returns a `GobReader`, which can be used to deserialize binary data. The `Read` function of the `GobReader` takes a pointer to the structure that the data should be written to.

The `NewGobWriterIO` function takes an `io.Writer` as a parameter and returns a `GobWriter`, which can be used to serialize Go structures into binary data. The `Write` function of the `GobWriter` takes a pointer to the structure that should be serialized.

Tests for these functions are also added to ensure that the encoding and decoding are working correctly.