feat(embedding): Change embedding and query size to 1536
BREAKING CHANGE: Change the size of embedding and query vectors to 1536 from the previous size of variable length float32 slices. This increases the accuracy of the embedding and improves the storage efficiency by reducing the length of the slice. The KeyValue struct also now uses a generic type for value.