feat(reflectutil): add IsZero function and tests, remove old zero function and tests
This commit adds a new IsZero function to the reflectutil package, which determines whether a given value is the zero value for its type. Tests are also provided for this new function. The old zero function and its tests are removed since they are no longer needed.

BREAKING CHANGE: The old zero function has been removed from the reflectutil package. Any code that was using this function will need to switch to using the new IsZero function instead.