feat(openai): update FetchEmbedding return type to fixed size embeddings
BREAKING CHANGE: The return type of FetchEmbedding in OpenaiTgenerator has been updated from [][]float32 to [][1536]float32 to accommodate fixed size embeddings. This is a breaking change and requires changes to any code that was using the old return type. Also, a new tokenizer has been added to OpenaiTgenerator to support multi-token phrases.