feat(tokenizer): add asynchronous initialization and encapsulation
This commit adds asynchronous initialization using singlewait package, to improve performance and encapsulate the tokenizer functionality within the Tokenizer struct. The tokenizer instance is now a member of the Tokenizer struct, which is instantiated using the NewTokenizer function. The CountTokens and OffsetTokens methods are now methods of the Tokenizer struct, while the RawTokens function remains a standalone function.