feat(redisembeddbconnector): change type of embedding parameter in AddEmbeddingDocument
The embedding parameter in the AddEmbeddingDocument function of RedisembedTgenerator has been changed from []float32 to [1536]float32 array, to match the expected size of the embeddings. This is a breaking change as it modifies the function signature, and may require changes in the client code calling this function.

BREAKING CHANGE: The AddEmbeddingDocument function in RedisembedTgenerator now expects an [1536]float32 array for the embedding parameter instead of a []float32 array.