feat: add FileInZip and unit test for WalkDirFromURL
This commit introduces a new implementation of types.FileReader interface called "FileInZip" for files that are contained in ZIP archives. It also adds a unit test case for `WalkDirFromURL()` which reads a zip file from the URL, lists all non-directory files inside the zip archive, ignores directories and returns `FileInZip` instances that can be used to retrieve the file contents. This feature expands the compatibility of tzap to also handle files and resources packaged as ZIP archives.