feat(cli): add version flag to display application version
This commit adds a Version variable to the cli/main.go file to hold the current version of the application. The version is set using a value passed in during the build process.

The commit also updates the main function to set the value of RootCmd.Version to the Version variable. This allows users to use the --version flag to display the current version of the application.

BREAKING CHANGE: None.