feat: update cosine_test.go to print filename from Metadata struct
This commit updates cosine_test.go to use the `Filename` field from the `Metadata` struct when printing the filename. It also removes commented-out code and changes the type of `results` to `float32`. Additionally, it updates test-files-1.json to have a newline at the end.

BREAKING CHANGE: The type of `results` in cosine_test.go has been changed from `[1536]float{}` to `[1536]float32{}`, which could potentially affect any code that uses this variable.