feat(cli): add reset command
This commit adds a new `reset` command to the CLI, which can be used to reset embeddings and other files. The command deletes the `.tzap-data/embeddingsCache.db`, `fileembeddings.db`, and `filesTimestamps.db` files if they exist. If the files do not exist, the command ignores them and continues.

The `reset` command is added as a subcommand under the root command.