fix(cli): use function NewQuery instead of GetQuery in LoadAndSearchEmbeddings
Previously, the `LoadAndSearchEmbeddings` function was using the `GetQuery` function from the `embed` package to retrieve the search query. This commit replaces it with the `NewQuery` function from the same package which is the intended function to use for creating new queries. This should fix the issue of failing tests related to searching embeddings.