feat: Add PrintSearchResults function to search for inspiration files
This commit adds a new function called PrintSearchResults to cli/cmd/cliworkflows/stdin.go that takes in a SearchResults struct and prints out the search results. The function is designed to search for inspiration files based on metadata. It loops through each result in the SearchResults struct and formats each vector using the cmdutil.FormatVectorToClickable function. It also counts tokens for each metadata split part using the CountTokens function. This new feature will improve the search functionality in the CLI by allowing users to search for inspiration files more easily.