fix(workflow): Fix typo and rename embedder loading error message
In the FetchOrCachedEmbeddingForFilesWorkflow function, there was a typo in the embedder loading error message. Additionally, the error message was changed to better reflect the actual error. The function now panics with an error if there is an issue loading the embedder. The function also calls the new renamed method `FetchThenCacheNewEmbeddings()` to fetch and cache new embeddings.