feat: refactor FileEvaluator to accept project name
This commit refactors the FileEvaluator struct to accept a project name in its constructor. This name is used to generate the absolute path to the project directory, which is then passed to the NewFilestampsDB function in the embed package. The updated New function in FileEvaluator is now also responsible for appending the name of the project directory to the .tzap-data directory name.

Additionally, the test for ShouldKeepPath has been moved to the fileinzip_test.go file and improved with helper functions.

BREAKING CHANGE: The NewFileEvaluator() function in cmdutil has been removed and replaced with the New() function in the fileevaluator package. Any dependencies using the old function will need to be updated to use the new one.