feat(cli): add option to search within a specific library
This commit adds the `--lib` flag to the search command, which can be used to specify a specific library to search within. The default value for `--lib` is an empty string, which indicates a search within the local project. If a library name is provided, the search will be restricted to files within that library.

BREAKING CHANGE: The `--zipurl` flag has been removed, as it was previously used to search within zip files. Searching within a specific library is now the recommended approach.