fix(install): handle error in case of invalid GitHub URL
Previously, the installation command would panic when an invalid GitHub URL was passed as an argument. This commit fixes the issue by handling the error and falling back to treating the argument as a direct URL to the ZIP file.

The code affected is in cli/cmd/install.go, where the GetZipUrlFromGithubUrl function is used to parse the input URL. If the function returns an error, the URL is assumed to be a direct link to the ZIP file, and the installation proceeds as usual.

This change improves the stability and user-friendliness of the installation process.