feat(cmdui): add support for watching file saves
This commit adds support for watching file saves in `cmdui` package. Now, when using `RunEditor()` function, if the editor is `"vscode"` or `"editor"`, `WatchSavesToFile()` function is called. This function watches the file for saves and if the file was saved `times` times within `changedWithin` duration, the watcher exits and returns `nil`.
This new feature enables the package to automatically re-prompt when the file is saved too many times within a set amount of time.