feat(cmdui): add MessageThread and refactor save/read to use MessageThread
This commit adds a new `MessageThread` data structure to `cmdui` package. `MessageThread` is used to represent a thread of messages with added functionality to check if the last message in the thread is from user and to get the last message. The `SaveThread`, `DeserializeThread` and `RunEditor` methods in `CMDUI` have been updated to use `SerializeMessageThread`, `ReadMessageThreadFromFile` and `SaveMessageThreadToFile` methods respectively, which operate on `MessageThread`.

BREAKING CHANGE: The `DeserializeThread` and `SaveThread` methods in `CMDUI` have been replaced with `ReadMessageThreadFromFile` and `SaveMessageThreadToFile` respectively. The signature of these methods have been changed to operate on `MessageThread` rather than `[]types.Message`.