feat(pkg/embed): improve performance of search by removing PCA
The current implementation of the search function in the `EmbedStore` package uses PCA for embedding vectors, which makes the loading of large projects slow. In this commit, PCA has been removed from the search function to improve performance. The PCA related code has been commented out and not deleted to allow easy rollback if required.

The `Stringutil` package has been updated to add functions for adding and stripping headers from embed files.