feat(openai): add retries for failed requests
BREAKING CHANGE: The OpenaiTgenerator struct now embeds the Tokenizer struct instead of having it as a field. This may cause compatibility issues for some clients that rely on the struct having a tokenizer field.

This commit adds the ability to retry failed requests in the OpenAI connector package. If a request fails due to a 429 status code (rate limit exceeded) or a 500 status code (server error), the client will wait for 2 seconds and retry the request. If the number of retries exceeds 3, the function will return an error.

The OpenaiTgenerator struct has been modified to embed the tokenizer struct instead of having it as a field. This may cause compatibility issues for some clients that rely on the struct having a tokenizer field.